proakad

+++ src/index.ts
@@ -21,7 +21,12 @@ program
 	.option("-o, --output <filepath>", "output root path")
 	.option("-d, --domain <domain>", "your app domain (eg. 'com.example.com')")
 	.option("-w, --overwrite", "overwrite existing files")
-	.option("-L, --lombok", "use lombok instead of getters and setters")
+	.option("-E, --entity", "generate entities")
+	.option("-S, --service", "generate services")
+	.option("-I, --serviceimpl", "generate service implementations")
+	.option("-C, --controller", "generate controllers")
+	.option("-R, --repository", "generate repositories")
+	.option("-F, --config", "generate config")
 	.parse(process.argv);
 console.log(filename);
 if (!fs.existsSync(filename)) {
@@ -76,7 +81,7 @@ jsonDDL.forEach(ent => {
 	if (isRelation(ent)) {
 		return;
 	}
-	const entity = new Entity(ent, domain, program.lombok);
+	const entity = new Entity(ent, domain, true);
 	const repository = new Repository(entity, domain);
 	const service = new Service(entity, domain);
 	const serviceImpl = new ServiceImpl(entity, domain);
@@ -86,15 +91,15 @@ jsonDDL.forEach(ent => {
 	const serviceImplFilename = join(serviceImplDir, serviceImpl.entity.className + "ServiceImpl.java");
 	const controllerFilename = join(controllerDir, entity.className + "Controller.java");
 	// @formatter:off
-	if (!fs.existsSync(entityFilename)      || program.overwrite) fs.writeFileSync(entityFilename, entity.toString());
-	if (!fs.existsSync(repositoryFilename)  || program.overwrite) fs.writeFileSync(repositoryFilename, repository.toString());
-	if (!fs.existsSync(serviceFilename)     || program.overwrite) fs.writeFileSync(serviceFilename, service.toString());
-	if (!fs.existsSync(serviceImplFilename) || program.overwrite) fs.writeFileSync(serviceImplFilename, serviceImpl.toString());
-	if (!fs.existsSync(controllerFilename)  || program.overwrite) fs.writeFileSync(controllerFilename, controllerTemplate(domain, entity));
+	if ((!fs.existsSync(entityFilename)      || program.overwrite) &&      program.entity) fs.writeFileSync(entityFilename, entity.toString());
+	if ((!fs.existsSync(repositoryFilename)  || program.overwrite) &&  program.repository) fs.writeFileSync(repositoryFilename, repository.toString());
+	if ((!fs.existsSync(serviceFilename)     || program.overwrite) &&     program.service) fs.writeFileSync(serviceFilename, service.toString());
+	if ((!fs.existsSync(serviceImplFilename) || program.overwrite) && program.serviceimpl) fs.writeFileSync(serviceImplFilename, serviceImpl.toString());
+	if ((!fs.existsSync(controllerFilename)  || program.overwrite) &&  program.controller) fs.writeFileSync(controllerFilename, controllerTemplate(domain, entity));
 	// @formatter:on
 });
-if (fs.existsSync(join(configDir, "Config.java")) && program.overwrite) {
+if (fs.existsSync(join(configDir, "Config.java")) && program.overwrite && program.config) {
 	fs.writeFileSync(join(configDir, "Config.java"), defaultConfig(domain));
-} else if (!fs.existsSync(join(configDir, "Config.java"))) {
+} else if (!fs.existsSync(join(configDir, "Config.java")) && program.config) {
 	fs.writeFileSync(join(configDir, "Config.java"), defaultConfig(domain));
 }
